Connecting back end and front end of weather agents. Implementing display of weather on the map

+++ backend/src/main/java/domain/extra/Weather.java
@@ -18,7 +18,7 @@ public class Weather extends Agent {
     private List<Route> orgRoutes = new ArrayList<>();
     private List<Route> altRoutes = new ArrayList<>();
 
-    public Weather(Coordinates initialCoordinates, double range) {
+    public Weather(Coordinates initialCoordinates, int range) {
         super(AgentType.WEATHER, initialCoordinates);
         super.kill();
         this.range = range;
